AnswerItem and AnswerTimeline refactored and tested, corrected imports in SubmitAnswerDialog for "dialog" imports
